new block and bettering of the current
new: can connect. which allows to check at runtime if a connection is
possible and act accordingly.
the current one has been adapted to give feedback that makes more sense
for beginners. also it turns off the card, as in the normal case, when
no wifi is working the whole program cannot work